Fixing review comments
